requirements.txt
PyQt6
python-barcode
pillow
